MangaUpdates tweaks (#7205)
* MangaUpdates tweaks

* Review changes


* Use built in functions instead of Jsoup

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/mangaupdates/MangaUpdatesApi.kt
@@ -125,7 +125,7 @@ class MangaUpdatesApi(
     suspend fun updateSeriesRating(track: Track) {
         if (track.score != 0f) {
             val body = buildJsonObject {
-                put("rating", track.score.toInt())
+                put("rating", track.score)
             }
             authClient.newCall(
                 PUT(
